requirements.txt
aiogram~=3.22.0
gspread~=6.2.1
python-dotenv~=1.1.1